requirements.txt
torch
torchvision
opencv-python
streamlit
numpy
pillow
